requirements.txt
absl-py                 ==       0.12.0
anyio                   ==      3.3.4
argon2-cffi             ==       20.1.0
asn1crypto              ==       0.24.0
astor                   ==       0.8.1
async-generator         ==       1.10
attrs                   ==       21.2.0
Babel                   ==       2.9.1
backcall                ==       0.2.0
bleach                  ==       3.3.0
Brlapi                  ==       0.6.7
cachetools              ==       4.2.2
certifi                 ==       2018.8.24
cffi                    ==       1.14.5
chardet                 ==       3.0.4
click                   ==       8.0.0
cryptography            ==       2.6.1
cupshelpers             ==       1.0
cycler                  ==       0.10.0
Cython                  ==       0.29.21
decorator               ==       4.3.0
defusedxml              ==       0.7.1
dill                    ==       0.2.8.2
distro                  ==       1.3.0
easydict                ==       1.9
entrypoints             ==       0.3
Flask                   ==       1.0.2
flatbuffers             ==       1.10
freetype-py             ==       2.2.0
ft2                     ==       1.0.2
future                  ==       0.18.2
gast                    ==       0.2.2
google-auth             ==       1.35.0
google-auth-oauthlib    ==       0.4.6
google-pasta            ==       0.2.0
gpio4                   ==       0.0.8
grpcio                  ==       1.37.1
h5py                    ==       2.8.0
idna                    ==       2.8
imageio                 ==       2.9.0
importlib-metadata      ==       4.0.1
ipykernel               ==       5.5.5
ipython                 ==       7.23.1
ipython-genutils        ==       0.2.0
ipywidgets              ==       7.6.3
itsdangerous            ==       2.0.1
jedi                    ==       0.17.2
Jinja2                  ==       3.0.1
joblib                  ==       1.0.0
json5                   ==       0.9.5
jsonschema              ==       3.2.0
jupyter-archive         ==       3.2.1
jupyter-client          ==       7.0.6
jupyter-core            ==       4.9.1
jupyter-server          ==       1.11.1
jupyterlab              ==       3.2.2
jupyterlab-language-pack-zh-CN == 3.2.post0
jupyterlab-pygments     ==       0.1.2
jupyterlab-server       ==       2.8.2
jupyterlab-widgets      ==       1.0.2
Keras                   ==       2.2.5
Keras-Applications      ==       1.0.8
Keras-Preprocessing     ==       1.1.2
keyring                 ==       17.1.1
keyrings.alt            ==       3.1.1
kiwisolver              ==       1.3.1
lmdb                    ==       0.93
louis                   ==       3.8.0
Markdown                ==       3.3.4
MarkupSafe              ==       2.0.1
matplotlib              ==       3.4.3
matplotlib-inline       ==       0.1.2
mistune                 ==       0.8.4
nbclassic               ==       0.2.8
nbclient                ==       0.5.3
nbconvert               ==       6.0.7
nbformat                ==       5.1.3
nest-asyncio            ==       1.5.1
networkx                ==       2.5
nlplateLPR              ==       1.1.0
notebook                ==       6.4.5
numpy                   ==       1.19.4
oauthlib                ==       3.1.1
olefile                 ==       0.46
onnx                    ==       1.6.0
onnx-tf                 ==       1.2.1
opt-einsum              ==       3.3.0
packaging               ==       20.9
paho-mqtt               ==       1.5.1
pandas                  ==       1.3.2
pandocfilters           ==       1.4.3
parso                   ==       0.7.1
peewee                  ==       3.14.4
pexpect                 ==       4.8.0
pickleshare             ==       0.7.5
Pillow                  ==       8.1.0
pip                     ==      21.1.1
ply                     ==       3.11
prometheus-client       ==       0.10.1
prompt-toolkit          ==       3.0.18
protobuf                ==       3.11.2
psutil                  ==       5.6.2
ptyprocess              ==       0.7.0
pyasn1                  ==       0.4.8
pyasn1-modules          ==      0.2.8
pycairo                 ==       1.16.2
pycparser               ==       2.20
pycrypto                ==       2.6.1
pycups                  ==       1.9.73
Pygments                ==       2.9.0
PyGObject               ==       3.30.4
pyparsing               ==       2.4.7
pyrsistent              ==       0.17.3
pyserial                ==       3.5
pysmbc                  ==       1.0.15.6
python-dateutil         ==      2.8.1
pytz                    ==       2021.1
PyWavelets              ==       1.1.1
pyxattr                 ==       0.6.1
pyxdg                   ==       0.25
PyYAML                  ==      5.3.1
pyzmq                   ==       22.0.3
requests                ==       2.22.0
requests-oauthlib       ==       1.3.0
requests-unixsocket     ==       0.2.0
rknn-toolkit            ==       1.7.1
RockX                   ==       1.2.0
rsa                     ==       4.7.2
ruamel.yaml             ==      0.15.81
scikit-image            ==       0.18.1
scikit-learn            ==       0.24.2
scipy                   ==       1.4.1
SecretStorage           ==       2.3.1
Send2Trash              ==       1.5.0
setuptools              ==      56.2.0
six                     ==     1.12.0
smbus2                  ==       0.4.0
sniffio                 ==     1.2.0
tb-mqtt-client          ==     1.2
tensorboard             ==     2.1.1
tensorflow              ==     2.1.0
tensorflow-estimator    ==     2.1.0
termcolor               ==     1.1.0
terminado               ==     0.9.5
testpath                ==     0.4.4
threadpoolctl           ==     2.1.0
tifffile                ==     2021.2.1
torch                   ==     1.6.0a0+b31f58d
tornado                 ==     6.1
traitlets               ==     5.0.5
typing-extensions       ==     3.10.0.0
uArm-Python-SDK         ==     2.0.6
urllib3                 ==     1.24.1
wcwidth                 ==     0.2.5
webencodings            ==     0.5.1
websocket-client        ==     0.59.0
Werkzeug                ==     2.0.1
wheel                   ==     0.36.2
widgetsnbextension      ==     3.5.1
wrapt                   ==     1.12.1
youtube-dl              ==     2019.1.17
zipp                    ==     3.4.1
rknn_toolkit-1.7.1-cp36-cp36m-linux_x86_64.whl